Fold SchedulingCoordinator's interface into its class
The interface had a single implementation beside it in the same package of
the same bundle, and the bundle exports neither, so it separated nothing.
The class now carries the documentation and declares itself as the service
it provides, which the generated component descriptor confirms.
